requirements.txt
bottle
python-vlc
fake-rpi
rpi.gpio
python-dateutil
